+++ tests/analysis/colocalization/test_prepare.py
@@ -27,7 +27,7 @@ def test_prepare_from_edgelist(edgelist):
     assert len(result.columns) == edgelist["marker"].nunique()
 
 
-@pytest.mark.parametrize("enable_backend", ["igraph", "networkx"], indirect=True)
+@pytest.mark.parametrize("enable_backend", ["networkx"], indirect=True)
 def test_prepare_from_graph(enable_backend, edgelist):
     graph = Graph.from_edgelist(
         edgelist=edgelist,
@@ -48,7 +48,7 @@ def test_prepare_from_graph(enable_backend, edgelist):
         assert len(result.columns) == len(unique_markers_in_component)
 
 
-@pytest.mark.parametrize("enable_backend", ["igraph", "networkx"], indirect=True)
+@pytest.mark.parametrize("enable_backend", ["networkx"], indirect=True)
 def test_prepare_from_graph_and_edgelist_eq_for_no_neigbours(enable_backend, edgelist):
     for _, component_df in edgelist.groupby("component", observed=True):
         graph = Graph.from_edgelist(