+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 * `graph` no longer outputs the raw edge list
+* igraph has been dropped at a graph backend for pixelator
 
 ## [0.15.2] - 2023-10-23